+++ assets/js/dashboard/App.vue
@@ -6,6 +6,9 @@
                 <router-view></router-view>
             </v-container>
         </v-content>
+        <v-overlay :value="$store.state.loading" color="white" opacity="0.7">
+            <v-progress-circular indeterminate size="64" color="teal accent-1"></v-progress-circular>
+        </v-overlay>
     </v-app>
 </template>
 
@@ -36,7 +39,9 @@ h1 {
 .theme--light.v-text-field > .v-input__control > .v-input__slot:before {
     border-color: rgba(0, 0, 0, 0.1);
 }
-.theme--light.v-text-field:not(.v-input--has-state) > .v-input__control > .v-input__slot:hover:before {
+.theme--light.v-text-field:not(.v-input--has-state)
+    > .v-input__control
+    > .v-input__slot:hover:before {
     border-color: rgba(0, 0, 0, 0.25);
 }
 </style>