+++ assets/js/dashboard/store/index.js
@@ -6,10 +6,29 @@ import gallery from "./modules/gallery";
 
 Vue.use(Vuex);
 
+const state = {
+    loading: false
+};
+
+const mutations = {
+    loading(state, payload) {
+        state.loading = payload;
+    }
+};
+
+const actions = {
+    loading({ commit }, payload) {
+        commit("loading", payload);
+    }
+};
+
 export default new Vuex.Store({
     modules: {
         users,
         profile,
         gallery
-    }
+    },
+    state,
+    mutations,
+    actions
 });